validation: service validation for autors, using custom exceptions and custom validators

+++ src/main/java/io/github/cursodsousa/libraryapi/service/AutorService.java
@@ -1,7 +1,10 @@
 package io.github.cursodsousa.libraryapi.service;
 
+import io.github.cursodsousa.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.cursodsousa.libraryapi.model.Autor;
 import io.github.cursodsousa.libraryapi.repository.AutorRepository;
+import io.github.cursodsousa.libraryapi.repository.LivroRepository;
+import io.github.cursodsousa.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,41 +15,59 @@ import java.util.UUID;
 public class AutorService {
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository){
+    public AutorService(
+            AutorRepository repository,
+            AutorValidator validator,
+            LivroRepository livroRepository
+    ) {
         this.repository = repository;
+        this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
-    public Autor salvar(Autor autor){
+    public Autor salvar(Autor autor) {
+        validator.validar(autor); // Validar com o validator que criamos (AutorValidator.validar())
         return repository.save(autor);
     }
 
-    public void atualizar(Autor autor){
-        if (autor.getId() == null){
+    public void atualizar(Autor autor) {
+        if (autor.getId() == null) {
             throw new IllegalArgumentException("Para atualizar, é necessário que o autor ja esteja salvo no database");
         }
+        validator.validar(autor); // Validar com o validator que criamos (AutorValidator.validar())
         repository.save(autor);
     }
 
-    public Optional<Autor> obterPorId(UUID id){
+    public Optional<Autor> obterPorId(UUID id) {
         return repository.findById(id);
     }
 
-    public void deletarPorId(Autor autor){
+    public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException("Nao é permitido excluir um autor que possui livros cadastrados!");
+        }
         repository.delete(autor);
     }
 
-    public List<Autor> pesquisa(String nome, String nacionalidade){
-        if (nome != null && nacionalidade != null){
+    public List<Autor> pesquisa(String nome, String nacionalidade) {
+        if (nome != null && nacionalidade != null) {
             return repository.findByNomeAndNacionalidade(nome, nacionalidade);
         }
-        if (nome != null){
+        if (nome != null) {
             return repository.findByNome(nome);
         }
-        if (nacionalidade != null){
+        if (nacionalidade != null) {
             return repository.findByNacionalidade(nacionalidade);
         }
 
         return repository.findAll();
+    }
+
+    // Verificar se autor tem algum livro
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
